Move Canvas xcm config to own module
